changing data type and request mapping

+++ server/src/main/java/com/admarch/model/Earnings.java
@@ -2,11 +2,12 @@ package com.admarch.model;
 
 public class Earnings {
     String registerNumber;
-    int conversionCount;
-    int conversionEarning;
+    long conversionCount;
+    long conversionEarning;
+
     public Earnings(String registerNumber,
-            int conversionCount,
-            int conversionEarning){
+            long conversionCount,
+            long conversionEarning){
         this.conversionCount = conversionCount;
         this.conversionEarning = conversionEarning;
         this.registerNumber =registerNumber;
