+++ server/src/main/java/com/admarch/model/LeaderBoard.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class LeaderBoard {
     String registerNumber;
-    int actionEarning;
-    public LeaderBoard(String registerNumber, int actionEarning){
+    long actionEarning;
+    public LeaderBoard(String registerNumber, long actionEarning){
         this.registerNumber = registerNumber;
         this.actionEarning = actionEarning;
     }
